Add a system property to keep the Java gui open.
The automatic closing of the window leads to program termination which
leads to some problems on *nix systems that run Xorg.
Per default anything copied to the clipboard by an application ceases to
be if the app is closed. Therefore the password copy feature is useless
in that case.

A system property `mpw.keep-gui` is introduced which will prevent the
automatic closing of the window if set to `1`.